# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Authenticates a user with email and password, issues an HttpOnly access cookie and an HttpOnly refresh cookie, and returns the authenticated user profile. In Swagger, this is the first endpoint to call before trying protected routes. The frontend should use `withCredentials` on subsequent requests instead of manually storing tokens.\n\nThe access token expires after 15m and the refresh token expires after 7d. The frontend should call `POST /auth/refresh` after a `401` caused by an expired access token.\n\n**Available credentials for the practice environment:**\n- Admin: `admin@example.com` / `admin123`\n- User: `user@example.com` / `user123`",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication completed successfully. Returns the user profile and sets the access and refresh cookies.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password."
          }
        },
        "summary": "Authenticate and issue access and refresh cookies",
        "tags": [
          "Access Management"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Uses the HttpOnly refresh cookie to issue a new short-lived access token and rotate the refresh token. In Swagger, call this after `POST /auth/login` if a protected endpoint starts returning 401 because the access cookie expired.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session renewed successfully. New access and refresh cookies were issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid refresh cookie."
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Rotate the refresh token and renew the access cookie",
        "tags": [
          "Access Management"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "description": "Returns the profile associated with the short-lived access cookie. In Swagger, use this endpoint immediately after `POST /auth/login` to verify that the browser session is already authenticated.",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated user profile returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access cookie."
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Retrieve the current authenticated user profile",
        "tags": [
          "Access Management"
        ]
      }
    },
...
    "/brands": {
      "get": {
        "operationId": "BrandsController_getAllBrands",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Vehicle brand catalog returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Brand"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List available vehicle brands",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/brands/{brandId}/models": {
      "get": {
        "operationId": "BrandsController_getModelsByBrand",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "description": "Brand identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model catalog for the requested brand returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Brand not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List models for a specific brand",
        "tags": [
          "Catalog"
        ]
      }
    }
  },
  "info": {
    "title": "NTT Demo Vehicle Catalog API",
    "description": "Backend API for the NTT Angular practice: authentication, vehicle catalog, brands, models, pagination and role-based access.\n\nRecommended flow in Swagger: call `POST /auth/login`, then use `GET /auth/me` and the protected endpoints normally. When the short-lived access cookie expires, call `POST /auth/refresh` and retry the failed request. Use `POST /auth/logout` to close the session cleanly.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "description": "Account password",
            "example": "admin123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier",
            "example": "1"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "admin@example.com"
          },
          "name": {
            "type": "string",
            "description": "Display name associated with the account",
            "example": "Admin User"
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the account. ADMIN can create, update and delete vehicles. USER has read-only access.",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "ADMIN"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "Profile of the authenticated user whose access and refresh cookies were issued",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            ]
          }
        },
        "required": [
          "user"
        ]
      },
...
      "Brand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Brand identifier",
            "example": "brand-1"
          },
          "name": {
            "type": "string",
            "description": "Display name of the brand",
            "example": "Toyota"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Model identifier",
            "example": "model-1"
          },
          "name": {
            "type": "string",
            "description": "Display name of the model",
            "example": "Corolla"
          },
          "brandId": {
            "type": "string",
            "description": "Identifier of the brand this model belongs to",
            "example": "brand-1"
          }
        },
        "required": [
          "id",
          "name",
          "brandId"
        ]
      }
    }
  }
}